% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonFunctions.R
\name{buildSampledDataList}
\alias{buildSampledDataList}
\title{buildSampledDataList}
\usage{
buildSampledDataList(fSet, controlMarkers = NULL, controlSize = 2000)
}
\arguments{
\item{controlSize}{}
}
\description{
buildSampledDataList
}

