% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{checkIntegrity}
\alias{checkIntegrity}
\title{Check integrity of data objects}
\usage{
checkIntegrity(annotation, data, mapVar, reconcile = TRUE)
}
\arguments{
\item{annotation}{- file manifest and annotations}

\item{data}{- data.table}

\item{reconcile}{- LOGICAL - reconcile ids in both annotation and data}
}
\value{

}
\description{
Check integrity of data objects
}
