% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotPlotModules.R
\name{dotPlot}
\alias{dotPlot}
\title{Title}
\usage{
dotPlot(data, aesList = aes_string(x = "Population", y = "percentPop"),
  xFacet, facetOrderList)
}
\arguments{
\item{xFacet}{}
}
\description{
Title
}
