% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotPlotModules.R
\name{dotPlotOutput}
\alias{dotPlotOutput}
\title{Percent Plot Module}
\usage{
dotPlotOutput(input, output, session, data, annotation, mapVar = c(name =
  "FCSFiles"), facetOrderList)
}
\arguments{
\item{data}{- populationTable}

\item{annotation}{- annotationTable (must be consistent with popTable)}
}
\description{
Percent Plot Module
}

