% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonFunctions.R
\name{findMedianValues}
\alias{findMedianValues}
\title{Find Median Values of Melted Data Frame}
\usage{
findMedianValues(controlMelt)
}
\arguments{
\item{controlMelt}{}
}
\description{
Find Median Values of Melted Data Frame
}
