% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonFunctions.R
\name{findPointsGeomTile}
\alias{findPointsGeomTile}
\title{Title}
\usage{
findPointsGeomTile(point, data, xcol, ycol, ps)
}
\arguments{
\item{ycol}{}
}
\value{
row in data that corresponds to clicked point
}
\description{
Title
}

