% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagSamples.R
\name{flagModuleOutput}
\alias{flagModuleOutput}
\title{Title}
\usage{
flagModuleOutput(input, output, session, data, annotation,
  idColumn = "patientID", subsetCondition = NULL, subsetChoices = NULL,
  sortConditions, markers, colorConditions, mapVar = c(idVar = "FCSFiles"))
}
\arguments{
\item{mapVar}{}
}
\description{
Title
}

