% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagSamples.R
\name{flagModuleUI}
\alias{flagModuleUI}
\title{Title}
\usage{
flagModuleUI(id, label = "qcViolin", markers, sortConditions, colorConditions,
  annotation)
}
\arguments{
\item{annotation}{}
}
\description{
Title
}

