% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingModules.R
\name{gatingModuleOutput}
\alias{gatingModuleOutput}
\title{Title}
\usage{
gatingModuleOutput(input, output, session, imageDir, popTable, displayNodes,
  annotation, plotObj, popSubsets, mapVar)
}
\arguments{
\item{input}{- shiny module input object}

\item{output}{- shiny module output object}

\item{session}{- shiny module session object}

\item{imageDir}{- relative directory location of "images/" directory generated
by plotAllPopulations()}

\item{popTable}{- populationTable generated by getPopulationsAndZscores}

\item{displayNodes}{- which gating populations to display in heatmap}

\item{annotation}{- annotation data.table}

\item{plotObj}{}
}
\description{
Title
}

