% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingModules.R
\name{gatingModuleOutputFromGO}
\alias{gatingModuleOutputFromGO}
\title{Title}
\usage{
gatingModuleOutputFromGO(input, output, session, GO, annotation, objId = NULL)
}
\arguments{
\item{plotObj}{}
}
\description{
Title
}

