% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingModules.R
\name{gatingModuleOutputGGFromGO}
\alias{gatingModuleOutputGGFromGO}
\title{Title}
\usage{
gatingModuleOutputGGFromGO(input, output, session, GO, annotation,
  objId = NULL)
}
\arguments{
\item{plotObj}{}
}
\description{
Title
}
