% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingModules.R
\name{gatingModuleUI}
\alias{gatingModuleUI}
\title{Title}
\usage{
gatingModuleUI(id, label = "gatingModule", popSubsets)
}
\arguments{
\item{id}{unique identifier for object}

\item{label}{unique name for UI module}

\item{annotation}{}
}
\description{
Title
}
