% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{gatingObj}
\alias{gatingObj}
\title{R6 object for Storing Gating Results and population percentages}
\format{\code{\link{qcFlowObj}} object.}
\usage{
gatingObj
}
\value{
Object of \code{\link{R6Class}} with methods for getting/setting features for flowDashboard.
}
\description{
R6 object for Storing Gating Results and population percentages
}
\details{
This object is generated from \code{\link{GOFromGatingSet}}. It is used in the following modules:
\code{gatingModule}, \code{dotPlotModule}, and \code{waterfallModule}.
}
\section{Fields}{

\describe{
\item{\code{qcData}}{\code{data.table} with three columns: \code{idVar}, \code{variable}, and \code{value}.
Usually generated by \code{QCOFromGatingSet}.}

\item{\code{annotation}}{Annotation (can be extracted as phenoData from a GatingSet) as \code{data.table}.}

\item{\code{subsetOptions}}{set which columns in annotation to use for subsetting. Set by \code{setSubsetAndSortOptions()}.}

\item{\code{subsetOptionList}}{named list, where every entry corresponds to levels in a column in annotation.}

\item{\code{}}{}
}}

\section{Methods}{

\describe{
  \item{\code{new()}}{}
  \item{\code{checkIntegrity()}}{This method checks whether the identifier used in annotation and data agree and ensures data integrity between the two.}
  \item{\code{setSubsetAndSortOptions()}}{set the subset and sortOptions}
  \item{\code{setMarkers()}}{set the markers in qcData and the markers to be displayed.}
  \item{\code{returnMergedData()}}{returns the merged qcData/annotation table as a data.table}
  }
}

\examples{
#Build a QCO Object from a GatingSet
gsFile <- system.file("extdata", "gvHDgs", package="flowDashboard")
gs <- load_gs(gsFile)
#default sampling is set to 4000 cells per FCS File
QCO <- QCOFromGatingSet(gs)
#look at slots of GatingObj
QCO

#build QCO from using new()
#Unless you want to build from scratch components, we recommend using
#QCOFromGatingSet

qc_data <- QCO$qcData
annot <- QCO$annotation
map_var <- QCO$mapVar
QCO <- qcFlowObj$new(qcData=qc_data, annotation=annot, mapVar = map_var)

}
\seealso{
\code{\link{QCOFromGatingSet}}
}
\keyword{data}
