% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{gatingObjFromGatingSet}
\alias{gatingObjFromGatingSet}
\title{Build a gatingObj from a gatingSet}
\usage{
gatingObjFromGatingSet(gs, annotation = NULL, populations = NULL,
  samplePop = 4000, imageDir = NULL, mapVar = NULL)
}
\arguments{
\item{gs}{- a GatingSet object.}

\item{annotation}{- annotation object. if NULL, will try to pull from phenoData in gs@data}

\item{mapVar}{- a named variable that will map to populationTable.
The column to map in popTable is `name`.}
}
\description{
Build a gatingObj from a gatingSet
}

