% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonFunctions.R
\name{makePopulationIdentifier}
\alias{makePopulationIdentifier}
\title{Title}
\usage{
makePopulationIdentifier(popName, name, pipelineFile = "Panel1",
  delimiter = "+")
}
\arguments{
\item{delimiter}{}
}
\description{
Title
}

