% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingModules.R
\name{popHeatmapGG}
\alias{popHeatmapGG}
\title{Title}
\usage{
popHeatmapGG(data, text = TRUE, xVar = NULL, yVar = NULL,
  fillVar = NULL, idVar = "idVar")
}
\arguments{
\item{mapVar}{}
}
\description{
Title
}
