% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{qcFlowObj}
\alias{qcFlowObj}
\title{R6 object for Quality Control of flow data}
\format{\code{\link{qcFlowObj}} object.}
\usage{
qcFlowObj
}
\value{
Object of \code{\link{R6Class}} with methods for getting/setting features for flowDashboard.
}
\description{
R6 object for Quality Control of flow data
}
\details{
This object is generated from \code{\link{QCOFromGatingSet}}. It is used in the following modules:
\code{qcModule} and \code{timeDriftModule}.
}
\section{Fields}{

\describe{
\item{\code{qcData}}{\code{data.table} with three columns: \code{idVar}, \code{variable}, and \code{value}.
Usually generated by \code{QCOFromGatingSet}.}

\item{\code{annotation}}{Annotation (can be extracted as phenoData from a GatingSet) as \code{data.table}.}

\item{\code{subsetOptions}}{set which columns in annotation to use for subsetting. Set by \code{setSubsetAndSortOptions()}.}

\item{\code{subsetOptionList}}{named list, where every entry corresponds to levels in a column in annotation.}

\item{\code{}}{}
}}

\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/lightning-viz/lightining-r/}
  \item{\code{new()}}{}
  \item{\code{checkIntegrity()}}{This method checks whether the identifier used in annotation and data agree and ensures data integrity between the two.}
  \item{\code{setSubsetAndSortOptions()}}{set the subset and sortOptions}
  \item{\code{setMarkers()}}{set the markers in qcData and the markers to be displayed.}
  \item{\code{returnMergedData()}}{returns the merged qcData/annotation table as a data.table}
  }
  @seealso \code{\link{QCOFromGatingSet}}
}

\examples{
library(flowWorkspace)
gs <- load_gs(flowStats::GvHD)

}
\keyword{data}
