% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCModules.R
\name{qcHeatmapGG}
\alias{qcHeatmapGG}
\title{Title}
\usage{
qcHeatmapGG(data, text = TRUE, xVar = "idVar", yVar = "variable",
  fillVar = "zscore", numVar = "med", lowColor = "blue",
  highColor = "orange")
}
\arguments{
\item{mapVar}{}
}
\description{
Title
}

