% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCModules.R
\name{qcHeatmapPlot}
\alias{qcHeatmapPlot}
\title{Title}
\usage{
qcHeatmapPlot(data, height = NULL, width = NULL, addText = TRUE,
  xVar = "idVar", yVar = "variable")
}
\arguments{
\item{data}{}
}
\description{
Title
}

