% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCModules.R
\name{qcModuleOutput}
\alias{qcModuleOutput}
\title{Title}
\usage{
qcModuleOutput(input, output, session, data, annotation, markers,
  colorConditions, mapVar = c(idVar = "FCSFiles"))
}
\arguments{
\item{annotation}{}

\item{idColumn}{}
}
\description{
Title
}
\examples{
Title

}

