% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCModules.R
\name{qcModuleUI}
\alias{qcModuleUI}
\title{Title}
\usage{
qcModuleUI(id, label = "qcViolin", markers, sortConditions, colorConditions,
  annotation)
}
\arguments{
\item{annotation}{}
}
\value{

}
\description{
Title
}
