% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{QCOFromGatingSet}
\alias{QCOFromGatingSet}
\title{Build a qcFlowObj from flowSet or gatingSet}
\usage{
QCOFromGatingSet(gs, annotation = NULL, samplePop = 4000,
  qcMarkers = NULL, mapVar = NULL, objId = NULL)
}
\arguments{
\item{gs}{- can be flowSet or gatingSet}

\item{annotation}{- annotation. if NULL, will attempt to get from phenoData slot}

\item{samplePop}{- number of points to sample from each flowFrame}

\item{qcMarkers}{- list of markers to return that represent qcMarkers}
}
\value{
qcFlowObj
}
\description{
Build a qcFlowObj from flowSet or gatingSet
}

