% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{reconcileData}
\alias{reconcileData}
\title{Reconcile annotation and data tables in object}
\usage{
reconcileData(annotation, data, mapVar, idsInBoth)
}
\arguments{
\item{idsInBoth}{}
}
\value{

}
\description{
Reconcile annotation and data tables in object
}
