% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonFunctions.R
\name{returnMeltedData}
\alias{returnMeltedData}
\title{Title}
\usage{
returnMeltedData(fS, selectMarkers = NULL, samplePop = NULL,
  returnCellNum = FALSE)
}
\arguments{
\item{fS}{- a flowSet.}

\item{selectMarkers}{- character vector of markers to Select. Must correspond to column names of flowFrame.}

\item{returnCellNum}{- Logical. return cell numbers (unique identifier for each row of flowFrame)}
}
\description{
Title
}

