% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonFunctions.R, R/fileManipulation.R
\name{returnMeltedDataFromGS}
\alias{returnMeltedDataFromGS}
\title{Title}
\usage{
returnMeltedDataFromGS(gS, population, removeMarkers = NULL,
  samplePopulation = NULL)

returnMeltedDataFromGS(gS, population, removeMarkers = NULL,
  samplePopulation = NULL)
}
\arguments{
\item{samplePopulation}{}

\item{samplePopulation}{}
}
\description{
Title

Title
}

