% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinModules.R
\name{violinOut}
\alias{violinOut}
\title{Title}
\usage{
violinOut(data, facets = NULL, colorVar = NULL, aggregateVar = NULL,
  population = NULL)
}
\arguments{
\item{aggregateVar}{}
}
\description{
Title
}

