% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinModules.R
\name{violinOutputFromPEO}
\alias{violinOutputFromPEO}
\title{Title}
\usage{
violinOutputFromPEO(input, output, PEO, annotation, aggregateList = NULL)
}
\arguments{
\item{annotation}{}
}
\description{
Title
}
