% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinModules.R
\name{violinUI}
\alias{violinUI}
\title{Title}
\usage{
violinUI(id, label = "qcViolin", markers, sortConditions, subsetCondition,
  populations, annotation)
}
\arguments{
\item{annotation}{}
}
\description{
Title
}

