% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfallModules.R
\name{waterfallOutput}
\alias{waterfallOutput}
\title{Waterfall plot output module}
\usage{
waterfallOutput(input, output, session, data, annotation, populationChoices,
  colorColumns, covariateChoices = NULL, subsetVariables = NULL,
  annotDisplayOptions = NULL, mapVar = c(name = "FCSFiles"))
}
\arguments{
\item{input}{- normal shiny input object}

\item{output}{- normal shiny output object}

\item{session}{- normal shiny session}

\item{subsetVariables}{}
}
\description{
Waterfall plot output module
}

