% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfallModules.R
\name{waterfallOutputFromGO}
\alias{waterfallOutputFromGO}
\title{Title}
\usage{
waterfallOutputFromGO(GO, annotation, objId = NULL)
}
\arguments{
\item{GO}{- a gating object}
}
\value{
- waterfallOutput module
}
\description{
Title
}
