% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfallModules.R
\name{waterfallOutputUIfromGO}
\alias{waterfallOutputUIfromGO}
\title{Title}
\usage{
waterfallOutputUIfromGO(GO, objId = NULL)
}
\arguments{
\item{GO}{- a `gatingObj`, built from \code{GOFromGatingSet}}
}
\description{
Title
}
