% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{commonDataObj}
\alias{commonDataObj}
\title{SuperClass for other DataObj Classes}
\format{\code{\link{R6Class}} object.}
\usage{
commonDataObj
}
\value{
Object of \code{\link{R6Class}} with methods for getting/setting features for flowDashboard.
}
\description{
SuperClass for other DataObj Classes
}
\section{Fields}{

\describe{
\item{\code{annotation}}{Annotation (can be extracted as phenoData from a GatingSet) as data.table.}

\item{\code{subsetOptions}}{set which columns in annotation to use for subsetting. Set by \code{setSubsetAndSortOptions()}.}
}}

\examples{
library(flowWorkspace)
gs <- load_gs(flowStats::GvHD)
#
}
\keyword{data}
