% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingModules.R
\name{gatingModuleUI}
\alias{gatingModuleUI}
\title{Title}
\usage{
gatingModuleUI(id, label = "qcViolin", sortConditions, subsetCondition,
  annotation)
}
\arguments{
\item{label}{}
}
\description{
Title
}

