% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingModules.R
\name{gatingModuleUI}
\alias{gatingModuleUI}
\title{Title}
\usage{
gatingModuleUI(id, label = "qcViolin", sortConditions,
  subsetCondition = NULL, annotation)
}
\arguments{
\item{annotation}{}
}
\description{
Title
}

