% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{PEOFromGatingSet}
\alias{PEOFromGatingSet}
\title{Build a populationExpressionObj from a GatingSet}
\usage{
PEOFromGatingSet(gs, annotation = NULL, populations = NULL,
  samplePop = 4000, objId = NULL)
}
\arguments{
\item{gs}{- a gatingSet}

\item{annotation}{- Annotations for each sample, where one column = sampleNames(gs).
If NULL, then it will attempt to grab annotation from the phenoData slot.}

\item{populations}{- A list of populations (must correspond to populationNames in gs).
If NULL, will just set populations with all populations in gatingSet}

\item{samplePop}{- Number of cells per population to sample. If NULL, returns all
cells in population.}
}
\value{
populationExpressionObj
}
\description{
Build a populationExpressionObj from a GatingSet
}

