% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingModules.R
\name{plotAllPopulations}
\alias{plotAllPopulations}
\title{Plot all Populations}
\usage{
plotAllPopulations(gateSet, nodeList, pipelineFile = "panel1",
  imagePath = "images/")
}
\arguments{
\item{gateSet}{- a gatingSet with attached populations}

\item{imagePath}{- directory to write population images}
}
\value{
nothing. Side effect is images written to the imagePath.
}
\description{
Given a gatingSet, plot the provenance for each sample and each population.
}

