% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{qcFlowObj}
\alias{qcFlowObj}
\title{R6 object for Quality Control of flow data}
\format{\code{\link{qcFlowObj}} object.}
\usage{
qcFlowObj
}
\value{
Object of \code{\link{R6Class}} with methods for getting/setting features for flowDashboard.
}
\description{
R6 object for Quality Control of flow data
}
\section{Fields}{

\describe{
\item{\code{annotation}}{Annotation (can be extracted as phenoData from a GatingSet) as data.table.}

\item{\code{subsetOptions}}{set which columns in annotation to use for subsetting. Set by \code{setSubsetAndSortOptions()}.}
}}

\examples{
library(flowWorkspace)
gs <- load_gs(flowStats::GvHD)
#
}
\keyword{data}
