% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCModules.R
\name{qcHeatmapGG}
\alias{qcHeatmapGG}
\title{Generates a QC heatmap using data generated by module}
\usage{
qcHeatmapGG(data, text = TRUE, xVar = "idVar", yVar = "variable",
  fillVar = "zscore", numVar = "med", lowColor = "blue",
  highColor = "gold")
}
\arguments{
\item{text}{Boolean. Should heatmap have}

\item{xVar}{- variable in data (should be factor) used for x-axis}

\item{yVar}{- variable in data (should be factor) used for y-axis}

\item{fillVar}{- variable in data used for fill for TCells}

\item{numVar}{-}

\item{highColor}{}
}
\description{
Generates a QC heatmap using data generated by module
}
\examples{
system.file("extdata", "")
}
