% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinModules.R
\name{violinOutput}
\alias{violinOutput}
\title{Title}
\usage{
violinOutput(input, output, session, data, annotation, facetList = NULL,
  aggregateList = NULL, markers = NULL, colorConditions = NULL,
  mapVar = c(idVar = "FCSFiles"))
}
\arguments{
\item{annotation}{}
}
\description{
Title
}
