% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfallModules.R
\name{waterfallOutput}
\alias{waterfallOutput}
\title{Title}
\usage{
waterfallOutput(input, output, session, data, annotation, populationChoices,
  colorColumns, subsetVariables = NULL)
}
\arguments{
\item{subsetVariables}{}
}
\description{
Title
}

