% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfallModules.R
\name{waterfallPlot}
\alias{waterfallPlot}
\title{Waterfall graphic expects a sorted table by population percentage}
\usage{
waterfallPlot(data, colorChoice)
}
\arguments{
\item{colorChoice}{}
}
\description{
Waterfall graphic expects a sorted table by population percentage
}
